[release/2023-10-30] Fix-forward: make swagger--clean now doesn't throw an error if run twice

+++ scripts/infrastructure/swagger.mk
@@ -7,8 +7,8 @@ SWAGGER_PUBLIC = $(SWAGGER_DIST)/public/swagger.yaml
 
 swagger--merge: $(SWAGGER_TIMESTAMP) ## Updates swagger builds from the components in the infrastructure/swagger/ directory.
 swagger--clean:  ## Removes swagger builds.
-	rm $(SWAGGER_TIMESTAMP)
-	rm -r $(SWAGGER_DIST)
+	[[ -f $(SWAGGER_TIMESTAMP) ]] && rm $(SWAGGER_TIMESTAMP) || :
+	[[ -d $(SWAGGER_DIST) ]] && rm -r $(SWAGGER_DIST) || :
 
 
 $(SWAGGER_TIMESTAMP): $(TIMESTAMP_DIR) $(SWAGGER_AWS) $(SWAGGER_PUBLIC)
